Rewrite Quaver parser.
Can actually represent the file it's loading now. Took a bit of work to figure out the exact flow of maintaining object depths and tracking but it works.

Honestly it was a nightmare to get working.

Also now generates a difficulty by using NPS that also factors in LNs, as those are generally harder.